Updated Packages; Replaced Google Places Package
Switching from react-google-places-autocomplete to
react-places-autocomplete.
Though the first component seems better and does not need the script tag
in the HTML to load the Places library, the second is what is used in
the lesson and is easier to make a component with since it is more
amenable to styling.