Set City and Venue Places Inputs in EventForm
Needed to fix the CORS problem because the API calls triggered by the
Venue Places Input were causing CORS errors.
Configured the City Places Input to show only cities.
Configured the Venue Places Input to use the value from the CIty input
to constrain the businesses that show up in the suggestions. Also
disabled the Venue input until a City was selected.