Added EventDetailsMap Component to EventDetailsInfo
Google-Map-React is NOT Typescript friendly. The rendering of markers
could be a lot better. Got it to work but it is still buggy; zooming in
and out causes the marker to drift. Not looking to fix it at the moment;
frankly I'd rather use a more TS friendly component or roll my own.