Refactored AsyncState to Be a Shared Type
So Neil, the author of the course, without the restrictions of
Typescript, seems to have made the choice to make async state shared
across the app by making an async reducer to manage when async methods
are loading (and blocking) or when those methods return errors. However,
Redux guidance is such that reducers should be pure functions with no
side effects. If one reducer is sending actions to the async reducer,
that's a side effect.
So I decided let every reducer with thunks manage its own async state.