Added Cloud Functions for Relationships
With the new security settings, users cannot change other users' data.
So Cloud Functions have to make changes on other profiles when the user
follows or unfollows other users.
CreateFollowerOnCreateFollowing creates a doc of the user as a follower
in the relationship doc of the target of their follow that triggers when the
user creates a doc in the relationships of the user they are following.